pyproject.toml
[tool.poetry]
name = "lilly-agent-eval"
version = "3.0.0"
description = "Simple, user-friendly agent evaluation platform"
authors = ["Eli Lilly and Company"]
readme = "README.md"
packages = [{include = "agent_eval"}]

[tool.poetry.dependencies]
python = "^3.11"

# Core
pydantic = "^2.6.0"

# Async HTTP
httpx = "^0.27.0"

# Web UI
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
python-multipart = "^0.0.9"

# YAML parsing
pyyaml = "^6.0.1"

# Templating
jinja2 = "^3.1.2"

# Environment
python-dotenv = "^1.0.0"

# File parsing
pypdf = "^4.0.1"
pdfplumber = "^0.11.0"
python-docx = "^1.1.0"
openpyxl = "^3.1.0"
pandas = "^2.2.0"

# LangSmith for execution tracing and observability
langsmith = "^0.7.0"

# OpenTelemetry for open-standard distributed tracing
opentelemetry-api = "^1.24.0"
opentelemetry-sdk = "^1.24.0"
opentelemetry-exporter-otlp = "^1.24.0"

# Optional: DeepEval for LLM-powered evaluations
deepeval = {version = "^1.2.0", optional = true}
openai = {version = "^1.40.0", optional = true}

# Optional: RAGAS for RAG-specific evaluation
ragas = {version = "^0.1.0", optional = true}
datasets = {version = "^2.14.0", optional = true}
langchain-openai = {version = "^0.1.0", optional = true}

# Optional: TruLens for feedback-based evaluation
trulens-eval = {version = "^0.30.0", optional = true}

# Optional: Statistical analysis
scipy = {version = "^1.13.0", optional = true}

[tool.poetry.extras]
deepeval = ["deepeval", "openai"]
ragas = ["ragas", "datasets", "langchain-openai"]
trulens = ["trulens-eval"]
all = ["deepeval", "openai", "ragas", "datasets", "langchain-openai", "trulens-eval", "scipy"]
stats = ["scipy"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
black = "^24.3.0"
ruff = "^0.3.0"

[tool.poetry.scripts]
agent-eval = "agent_eval.cli:main"
start-eval = "agent_eval.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
